requirements.txt
aiogram==3.26.0
python-dotenv
sqlalchemy
asyncpg
alembic
